Face-Detection-Tracking/requirements.txt
matplotlib==3.3.2
mtcnn==0.1.0
scikit_image==0.17.2
numpy==1.19.5
streamlit==0.79.0
numba==0.53.1
filterpy==1.4.5
opencv_python_headless==4.5.1.48
Pillow==8.2.0
scikit_learn==0.24.2
skimage==0.0
